Create an event listener to restart game
Also make the variables and methods of the ShipPlacementEvent class
instance variables and methods instead of static so that with each new
game, a new instance of the events could be used to start the game.

Also, put the code that is activated when the player attacks in a
seperate function and pass it into the event listener. This is done so
the event listener listening to this function could be removed when the
game ends.